Move `FilesViewItemView` to its own file
